Delete webboots pkg/dhclient to use u-root's pkg/dhclient

removed kernel panic by getting result of interface
